requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.3.0
Django==2.2.7
django-ckeditor==5.8.0
django-crispy-forms==1.8.0
django-js-asset==1.2.2
importlib-metadata==0.23
kombu==4.6.6
more-itertools==7.2.0
Pillow==6.2.1
python-slugify==4.0.0
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0
text-unidecode==1.3
vine==1.3.0
zipp==0.6.0
